📝 Job Page - Add React-Router and Search Page
